Renamed RecipeTypeSorted to RecipeTypeProcessingOrdered
