+++ src/main/java/thelm/packagedauto/proxy/CommonProxy.java
@@ -22,7 +22,7 @@ import thelm.packagedauto.network.GuiHandler;
 import thelm.packagedauto.network.PacketHandler;
 import thelm.packagedauto.recipe.RecipeTypeCrafting;
 import thelm.packagedauto.recipe.RecipeTypeProcessing;
-import thelm.packagedauto.recipe.RecipeTypeSorted;
+import thelm.packagedauto.recipe.RecipeTypeProcessingOrdered;
 import thelm.packagedauto.tile.TileCrafter;
 import thelm.packagedauto.tile.TileEncoder;
 import thelm.packagedauto.tile.TilePackager;
@@ -92,7 +92,7 @@ public class CommonProxy {
 
 	protected void registerRecipeTypes() {
 		RecipeTypeRegistry.registerRecipeType(RecipeTypeProcessing.INSTANCE);
-		RecipeTypeRegistry.registerRecipeType(RecipeTypeSorted.INSTANCE);
+		RecipeTypeRegistry.registerRecipeType(RecipeTypeProcessingOrdered.INSTANCE);
 		if(TileCrafter.enabled) {
 			RecipeTypeRegistry.registerRecipeType(RecipeTypeCrafting.INSTANCE);
 		}
